+++ lib/hausgold/entity/property.rb
@@ -9,6 +9,7 @@ module Hausgold
 
     # Mapped and tracked attributes
     tracked_attr :id, :owner_id, :permissions, :object_details, :lead_id,
-                 :metadata, :source, :created_at, :updated_at, :geo_details
+                 :metadata, :source, :created_at, :updated_at, :valuation_at,
+                 :geo_details, :valuation
   end
 end