+++ lib/hausgold/entity/search_profile.rb
@@ -9,8 +9,8 @@ module Hausgold
 
     # Mapped and tracked attributes
     tracked_attr :id, :gid, :user_id, :usages, :property_types,
-                 :property_subtypes, :city, :zipcode, :perimeter, :price_from,
-                 :price_to, :year_of_construction_from,
+                 :property_subtypes, :city, :zipcode, :perimeter,
+                 :currency, :price_from, :price_to, :year_of_construction_from,
                  :year_of_construction_to, :amount_rooms_from,
                  :amount_rooms_to, :living_space_from, :living_space_to,
                  :land_size_from, :land_size_to, :created_at, :updated_at